test: revamp e2e test setup, add user journey tests
In preparation for adding some user journey end-to-end integration
tests, this replaces the existing e2e test setup with one that will be
better suited for those needs.

Specifically, instead of using a GitHub Action workflow for each fixture that is
responsible for waiting for the deployment of that fixture - which was triggered by a
Netlify site which has been linked to this repo - we simply write tests, and any given
test can depend on any given fixture, which is built and deployed lazily on demand. This
allows us to (1) avoid needing to write a CI workflow for each fixture, (2) avoid needing
to couple behaviors under test to the test setup, and (3) follows a more explicit pattern
for explicitly running these tests again the code in this repo (as opposed to `latest`
published).

I also changed the "demos" (renamed "fixtures") to no longer be monorepo packages within
this repo. Since handling monorepos is possibly within the scope of complexities handled
by our adapters, it seems wise not to blur the lines here by having a monorepo package
tested as a non-monorepo site. The setup was difficult to wrangle anyway.