ci: fix Remix bump failure notifier workflow (#365)
I couldn't quite figure out why, but the `check_suite` events weren't firing.

I couldn't use the `status` events, because those are only about the legacy(ish)
Commit Status API, which is completely separate from the Checks API, which all
GitHub Apps and all GitHub Actions use.

I preferred not to use the `workflow_run` events, because it requires explicitly
listing the other workflows it depends on - it doesn't support listening to all
workflow events. I didn't like the brittle action-at-a-distance created by
needing to keep the full list of workflows in sync between this file and other
config locations.

I couldn't trivially use the `pull_request` events, because there is no event
that triggers when the PR (or HEAD commit) status(es) change(s) - i.e. at the
time of the PR opening or being pushed to, the checks haven't actually completed
yet. It seemed my best option was to have this workflow block and wait for all
other workflows to resolve. Luckily, I found an existing open-source Action for
this. This workflow is more complex than I'd like it to be, but it's the best we
can do, I think.