feat(remix-edge-adapter): use site entrypoint for Hydrogen sites
It appears to be (nearly?) imposssible to provide an out-of-the-box Netlify
experience for Hydrogen sites that use Remix Vite. We may be able to
solve for this at some point, but for now our only option is to expect
these sites to contain a `server.ts` (or similar) file at the root. This
is what Hydrogen templates do and what the Netlify Hydrogen template
will do.

There may be value in supporting this for Remix sites as well, but I
haven't heard of a use case so I didn't bother supporting it here.