feat(@netlify/vite-plugin-react-router)!: support React Router middleware (#546)
Ultimately, the essence of this change is just to pass an instance of `ReactContextProvider` to the
RR request handler; exporting a `netlifyContextProvider` additionally allows access to Netlify
context in middleware, and the rest is docs and test changes.

It's too much of a hassle to support all the combinations of scenarios if we support earlier
versions where the RR `createContext` and `RouterContextProvider` exports don't exist at all and so
on. Instead, this makes a breaking change that requires 7.9.0+.

fix(deps): move react-router to dev deps

fix(deps): remove unused @react-router/node dep

---------

Co-authored-by: Michal Piechowiak <misiek.piechowiak@gmail.com>